chore: remove old event detail page
