+++ app/pages/events/[id].vue
@@ -1,37 +1,107 @@
-<script setup lang="ts">
+<script setup>
 import 'maplibre-gl/dist/maplibre-gl.css';
-import { ref, computed, onMounted } from "vue";
+import { ref, computed, onMounted } from "vue"
 
 const route = useRoute()
+
+// Get the ID from the route params
+const eventId = route.params.id
 const { data: event, error } = await useFetch(`/api/events/${route.params.id}`)
 
+const loading = ref(true)
+const notFound = ref(false)
+
 const isEditMode = ref(false)
-const editForm = ref({ ...event.value })
+const editForm = ref({})
 
+//placeholder until we implement auth
+const admin = true;
 
-if (error.value) {
-  console.error(error.value)
-  showError({
-    statusCode: error.value.statusCode,
-    statusMessage: "The requested event could not be fetched."
+const style = '/mapstyles.json'
+const center = [-96.77049780046936, 32.772891246510596]
+const zoom = 15
+
+// Fetch event data from API on mount
+onMounted(async () => {
+  try {
+    console.log('📡 Loading event with ID:', eventId)
+    
+    // Fetch event from your backend API
+    event.value = await $fetch(`/api/events/${eventId}`)
+    
+    editForm.value = { ...event.value }
+    
+    console.log('✅ Event loaded:', event.value)
+    loading.value = false
+    
+  } catch (error) {
+    console.error('❌ Error fetching event:', error)
+    loading.value = false
+    notFound.value = true
+  }
+})
+
+const formattedDate = computed(() => {
+  if (!event.value) return ''
+  
+  const start = new Date(event.value.startTime)
+  const end = new Date(event.value.endTime)
+  
+  const dateStr = start.toLocaleDateString('en-US', { 
+    weekday: 'long', 
+    year: 'numeric', 
+    month: 'long', 
+    day: 'numeric' 
   })
+  
+  const startTime = start.toLocaleTimeString('en-US', { 
+    hour: 'numeric', 
+    minute: '2-digit' 
+  })
+  
+  const endTime = end.toLocaleTimeString('en-US', { 
+    hour: 'numeric', 
+    minute: '2-digit' 
+  })
+  
+  return `${dateStr} • ${startTime} - ${endTime}`
+})
+
+function toggleEditMode() {
+  if (isEditMode.value) {
+    // Cancel editing - reset form to original event data
+    editForm.value = { ...event.value }
+  }
+  isEditMode.value = !isEditMode.value
 }
 
-const toast = useToast()
-
-const showToast = (success: boolean) => {
-  if(success) {
-    toast.add({
-      title: "Successfully uploaded image.",
-      icon: "i-lucide-image-up",
-      color: "success"
+async function saveChanges() {
+  try {
+    console.log('💾 Saving changes...')
+    
+    // Update event via API
+    await $fetch(`/api/events/${event.value.id}`, {
+      method: 'PATCH',
+      body: {
+        title: editForm.value.title,
+        shortDesc: editForm.value.shortDesc,
+        description: editForm.value.description,
+        location: editForm.value.location,
+        startTime: new Date(editForm.value.startTime).toISOString(),
+        endTime: new Date(editForm.value.endTime).toISOString(),
+        allowVolunteers: editForm.value.allowVolunteers,
+        allowAttendees: editForm.value.allowAttendees
+      }
     })
-  } else {
-    toast.add({
-      title: "Error uploading image.",
-      icon: "i-lucide-circle-x",
-      color: "error"
-    })
+    
+    // Update local event data
+    event.value = { ...editForm.value }
+    isEditMode.value = false
+    
+    console.log('✅ Event updated successfully')
+  } catch (error) {
+    console.error('❌ Error updating event:', error)
+    alert('Error updating event. Please try again.')
   }
 }
 
@@ -44,76 +114,162 @@ const placeholders = [
   'https://picsum.photos/640/640?random=6'
 ]
 
+const filesToUpload = ref([])
+
+function handleImageUpload(event) {
+  const files = event.target?.files || event
+  
+  if (!files || files.length === 0) return
+  
+  // Store actual File objects for upload
+  filesToUpload.value = [...filesToUpload.value, ...Array.from(files)]
+  
+  // Create preview URLs
+  const imagePromises = Array.from(files).map(file => {
+    return new Promise((resolve) => {
+      const reader = new FileReader()
+      reader.onload = (e) => resolve({ 
+        imageUrl: e.target.result,
+        fileName: file.name,
+        isPreview: true 
+      })
+      reader.readAsDataURL(file)
+    })
+  })
+  
+  Promise.all(imagePromises).then(images => {
+    editForm.value.eventAssets = [...(editForm.value.eventAssets || []), ...images]
+  })
+}
+
+async function uploadNewImages() {
+  if (filesToUpload.value.length === 0) return
+  
+  console.log(`📤 Uploading ${filesToUpload.value.length} new images...`)
+  
+  for (const file of filesToUpload.value) {
+    const formData = new FormData()
+    formData.append('file', file)
+    
+    try {
+      await $fetch(`/api/events/${event.value.id}/images/upload`, {
+        method: 'POST',
+        body: formData
+      })
+      console.log(`✅ Uploaded: ${file.name}`)
+    } catch (uploadError) {
+      console.error(`❌ Failed to upload ${file.name}:`, uploadError)
+    }
+  }
+  
+  // Clear the upload queue
+  filesToUpload.value = []
+  
+  // Reload event to get updated image URLs
+  const updatedEvent = await $fetch(`/api/events/${event.value.id}`)
+  event.value = updatedEvent
+  editForm.value = { ...updatedEvent }
+}
+
+function removeImage(index) {
+  const asset = editForm.value.eventAssets[index]
+  
+  // If it's a preview (not yet uploaded), just remove from array
+  editForm.value.eventAssets.splice(index, 1)
+  filesToUpload.value.splice(index, 1)
+
+}
+
+// Helper to get proper image URL
+function getImageUrl(asset) {
+  if (asset.isPreview) {
+    return asset.imageUrl // Base64 preview
+  }
+  // Use your backend image endpoint
+  return `/api/events/${event.value.id}/images/${asset.imageUrl.split('/').pop()}`
+}
+
 const fetchedItems = event.value?.eventAssets.map((asset) => "/api/events/" + asset.imageUrl)
 
 const items = (fetchedItems?.length || 0) > 0 ? fetchedItems : placeholders
-
-const style = '/mapstyles.json'
-const center = [-96.77049780046936, 32.772891246510596]
-const zoom = 15
-
-const fileUpload = async (file: File | null | undefined) => {
-  if(!file) {
-    return
-  }
-
-  const formData = new FormData()
-  formData.append("file", file)
-
-  const res = await $fetch.raw(`/api/events/${route.params.id}/images/upload`, {
-    method: 'POST',
-    body: formData,
-    ignoreResponseError: true
-  })
-
-  showToast(res.status === 201)
-}
-
-
-
-const admin = true;
-
 </script>
 
-
 <template>
-  <div v-if="!event">Loading...</div>
-  <div v-else>
-    <div class="flex flex-col mt-12 mb-12 w-full h-full bg-gray-100 items-center">
+  <div class="min-h-screen bg-gray-50 pb-24">
+    <!-- Loading State -->
+    <div v-if="loading" class="flex items-center justify-center min-h-screen">
+      <div class="text-center">
+        <div class="animate-spin rounded-full h-12 w-12 border-b-2 border-primary-600 mx-auto mb-4"/>
+        <p class="text-gray-600">Loading event...</p>
+      </div>
+    </div>
 
-      <div class="flex gap-2 justify-left w-full max-w-4xl px-4 mt-6">
+    <!-- Not Found State -->
+    <div v-else-if="notFound" class="flex items-center justify-center min-h-screen">
+      <div class="text-center">
+        <UIcon name="i-lucide-calendar-x" class="w-16 h-16 text-gray-400 mx-auto mb-4" />
+        <h2 class="text-2xl font-bold text-gray-900 mb-2">Event Not Found</h2>
+        <p class="text-gray-600 mb-4">The event you're looking for doesn't exist.</p>
+        <UButton
+          icon="i-lucide-arrow-left"
+          @click="navigateTo('/eventManagement')"
+        >
+          Back to Events
+        </UButton>
+      </div>
+    </div>
+
+    <!-- Event Details -->
+    <div v-else>
+      <!-- Header with Edit Button -->
+      <div class="bg-white shadow-sm sticky top-0 z-10 mt-16">
+        <div class="max-w-4xl mx-auto px-4 py-4 flex items-center justify-between">
+          <UButton
+            icon="i-lucide-arrow-left"
+            variant="ghost"
+            class="text-brand4"
+            @click="navigateTo('/events/manage')"
+          />
+          
+          <div v-if="admin"class="flex gap-2">
             <UButton
-              v-if="!isEditMode && admin"
+              v-if="!isEditMode"
               icon="i-lucide-pencil"
               color="brand4"
               variant="soft"
-              @click="isEditMode = true"
+              @click="toggleEditMode"
             >
               Edit Event
             </UButton>
             
-            <template v-if="isEditMode">
+            <template v-else>
               <UButton
                 variant="ghost"
                 color="brand4"
-                @click="isEditMode = false"
+                @click="toggleEditMode"
               >
                 Cancel
               </UButton>
               <UButton
                 icon="i-lucide-check"
                 color="brand4"
+                @click="async () => { await saveChanges(); await uploadNewImages(); }"
               >
                 Save Changes
               </UButton>
             </template>
           </div>
-          
+        </div>
+      </div>
 
-      <UContainer class="flex-1 w-full h-full overflow-y-auto py-10">
-         
-        <h1  v-if="!isEditMode"  class="text-3xl font-hornbill font-bold mb-2 text-brand4 text-center">{{ event.title }}</h1>
-        <UInput
+      <div class="max-w-4xl mx-auto px-4 py-8">
+        
+        
+        <!-- Event Title -->
+        <div class="mb-6">
+          <h1  v-if="!isEditMode"  class="text-3xl font-hornbill font-bold mb-2 text-brand4 text-center">{{ event.title }}</h1>
+
+          <UInput
             v-else
             v-model="editForm.title"
             size="xl"
@@ -121,19 +277,30 @@ const admin = true;
             class="text-4xl font-bold"
             
           />
-        <p v-if="!isEditMode" class="text-gray-600 text-xl font-poppins mb-4 text-center">{{ event.shortDesc }}</p>
-        <UInput
+        </div>
+
+        <!-- Short Description -->
+        <div v-if="event.shortDesc || isEditMode" class="bg-brand6 rounded-2xl p-3 mb-6">
+          <p v-if="!isEditMode" class="text-md text-gray-700 italic">
+            {{ event.shortDesc }}
+          </p>
+          <UInput
             v-else
             v-model="editForm.shortDesc"
-            size="xl"
-            placeholder="Event Title"
-            class="text-4xl font-bold"
-            
+            class="w-full"
+            placeholder="Short Description"
+            size="lg"
           />
-        <UCarousel v-if="!isEditMode" v-slot="{ item }" dots :items="items" class="h-80 max-w-xs mx-auto">
+        </div>
+
+        <!-- Event Images Carousel -->
+        <div class="mb-8">
+          <UCarousel v-if="!isEditMode" v-slot="{ item }" dots :items="items" class="h-80 max-w-xs mx-auto">
           <img :src="item" class="h-80 w-auto rounded-lg mx-auto">
-        </UCarousel>
-        <div v-else class="space-y-4">
+          </UCarousel>
+          
+          <!-- Edit Mode Images -->
+          <div v-else class="space-y-4">
             <div v-if="editForm.eventAssets && editForm.eventAssets.length > 0" class="grid grid-cols-2 gap-4">
               <div 
                 v-for="(asset, index) in editForm.eventAssets"
@@ -174,9 +341,71 @@ const admin = true;
               >
             </div>
           </div>
-        <div class="flex-1 mt-4 mb-10 w-full h-full overflow-y-auto"/>
+        </div>
+        
 
-        <p class="text-gray-600 font-poppins mb-4">{{ event.description }}</p>
+        <!-- Date & Location Card -->
+        <div class="bg-white rounded-2xl shadow-sm p-6 mb-6">
+          <div class="flex items-start gap-4 mb-4">
+            <div class="bg-brand6 p-3 rounded-xl">
+              <UIcon name="i-lucide-calendar" class="w-6 h-6 text-brand4" />
+            </div>
+            <div class="flex-1">
+              <p class="text-sm text-gray-500 mb-1">Date & Time</p>
+              <p v-if="!isEditMode" class="text-gray-900 font-medium">
+                {{ formattedDate }}
+              </p>
+              <div v-else class="grid grid-rows-2 gap-2">
+                <div>
+                  <label class="text-xs text-gray-500">Start </label>
+                  <UInput v-model="editForm.startTime" type="datetime-local" />
+                </div>
+                <div>
+                  <label class="text-xs text-gray-500">End </label>
+                  <UInput v-model="editForm.endTime" type="datetime-local" />
+                </div>
+              </div>
+            </div>
+          </div>
+
+          
+
+          <div class="flex items-start gap-4">
+            <div class="bg-brand6 p-3 rounded-xl">
+              <UIcon name="i-lucide-map-pin" class="w-6 h-6 text-brand4" />
+            </div>
+            <div class="flex-1">
+              <p class="text-sm text-gray-500 mb-1">Location</p>
+              <p v-if="!isEditMode" class="text-gray-900 font-medium">
+                {{ event.location }}
+              </p>
+              <UInput
+                v-else
+                v-model="editForm.location"
+                placeholder="Event Location"
+              />
+            </div>
+          </div>
+        </div>
+        
+        <div v-if="!isEditMode" id="map" class="h-60 relative overflow-hidden justify-center items-center mb-6">
+          <MapInteractive :style="style" :center="center" :zoom="zoom" />
+        </div>
+
+        <!-- Full Description -->
+        <div class="bg-white rounded-2xl shadow-sm p-6 mb-6">
+          <h2 class="text-2xl font-semibold mb-4">About This Event</h2>
+          <p v-if="!isEditMode" class="text-gray-700 leading-relaxed whitespace-pre-line">
+            {{ event.description }}
+          </p>
+          <UTextarea
+            v-else
+            v-model="editForm.description"
+            class="w-full"
+            placeholder="Full event description"
+            :rows="6"
+          />
+        </div>
 
         <div v-if="event.mobileClinicId" class="mt-4 mb-4">
           <p class="text-gray-600 font-poppins">
@@ -184,57 +413,67 @@ const admin = true;
           </p>  
         </div>
 
-        <p class="text-gray-600 font-poppins mb-4">
-          📍 {{ event.location }}  
-          <br >
-          🕒 {{ new Date(event.startTime).toLocaleString() }} → {{ new Date(event.endTime).toLocaleString() }}
-        </p>
-        
-        <div id="map" class="h-60 relative overflow-hidden justify-center items-center">
-          <MapInteractive :style="style" :center="center" :zoom="zoom" />
-        </div>
-
-        <div class="flex-1 mt-4 mb-2 w-full h-full overflow-y-auto"/>
-
-        <UContainer>
-          <div class="mt-4 text-center">
-            <div class="flex gap-4 justify-center">
-              <div v-if="event.allowVolunteers" class="flex-1 max-w-xs px-2">
-                <UButton
-                  class="w-full bg-brand4 text-white font-poppins rounded-xl py-3"
-                  :ui="{ base: 'justify-center' }"
-                  size="xl"
-                  href="/volunteer"
-                >
-                  Volunteer
-                </UButton>
+        <!-- Event Options -->
+        <div v-if="admin && isEditMode" class="bg-white rounded-2xl shadow-sm p-6 mb-6">
+          <h2 class="text-2xl font-semibold mb-4">Event Settings</h2>
+          
+          <div class="space-y-4">
+            <div class="flex items-center justify-between p-4 bg-gray-50 rounded-xl">
+              <div class="flex items-center gap-3">
+                <UIcon name="i-lucide-users" class="w-5 h-5 text-brand4" />
+                <div>
+                  <p class="font-medium text-gray-900">Volunteer Sign-ups</p>
+                  <p class="text-sm text-gray-500">Allow people to volunteer for this event</p>
+                </div>
               </div>
-              <div v-if="event.allowAttendees" class="flex-1 max-w-xs px-2">
-                <UButton
-                  class="w-full bg-brand4 text-white font-poppins rounded-xl py-3"
-                  :ui="{ base: 'justify-center' }"
-                  size="xl"
-                  :href="`/events/${event.id}/register`"
-                >
-                  Register
-                </UButton>
+              <label class="flex items-center gap-2 cursor-pointer">
+                <UCheckbox v-model="editForm.allowVolunteers" color="brand4" />
+              </label>
+            </div>
+
+            <div class="flex items-center justify-between p-4 bg-gray-50 rounded-xl">
+              <div class="flex items-center gap-3">
+                <UIcon name="i-lucide-ticket" class="w-5 h-5 text-brand4" />
+                <div>
+                  <p class="font-medium text-gray-900">Attendee Registration</p>
+                  <p class="text-sm text-gray-500">Allow people to register as attendees</p>
+                </div>
               </div>
+              <label v-if="isEditMode" class="flex items-center gap-2 cursor-pointer">
+                <UCheckbox v-model="editForm.allowAttendees" color="brand4" />
+              </label>
+              <label v-else class="flex items-center gap-2">
+                <UCheckbox :model-value="event.allowAttendees" color="brand4" disabled />
+              </label>
             </div>
           </div>
-          <div class="h-12" />
-          <UFileUpload 
-            accept="image/*" 
-            label="Upload an Image for this event!"
-            description="Image files only"
-            @update:model-value="fileUpload"/>
-        </UContainer>
+        </div>
 
-        <div class="flex-1 mt-4 mb-8 w-full h-full overflow-y-auto"/>
-      
-      </UContainer>    
-      
-      
+        <!-- Action Buttons for Attendees/Volunteers -->
+        <div v-if="!isEditMode" class="flex gap-4">
+          <UButton
+            v-if="event.allowVolunteers"
+            color="brand4"
+            size="xl"
+            block
+            icon="i-lucide-heart-handshake"
+          >
+            Sign Up as Volunteer
+          </UButton>
           
+          <UButton
+            v-if="event.allowAttendees"
+            color="brand4"
+            variant="outline"
+            size="xl"
+            block
+            icon="i-lucide-ticket"
+          >
+            Register to Attend
+          </UButton>
+        </div>
+
+      </div>
     </div>
   </div>
-</template>    
+</template>
